Fix initial make error, add env rebuild script, add dxf parser, fix benchmark.

+++ rebuild-env.py
@@ -2,38 +2,46 @@
 #
 # Rebuilds virtualenv after pip changes.
 
-# TODO(igorc): Make this script pass rather than first fail
-# and then having to do "make develop" and move the clips.
-# It also seems to do poor job saving/restoring clips,
-# as player complains about missing playlists.
-
 import os
 import shutil
 import subprocess
 import sys
 
-_CLIPS_DST = 'env/clips'
-_CLIPS_BAK = 'clips_bak'
+def move_dir(src_dir, dst_dir):
+  if not os.path.exists(src_dir):
+    return True
+  if os.path.exists(dst_dir):
+    print 'ERROR: Cannot overwrite:', dst_dir
+    return False
+  os.rename(src_dir, dst_dir)
+  return True
+
+
+def backup(dir, reverse):
+  src_dir = 'env/' + dir
+  backup_dir = '_backup_' + dir
+  if reverse:
+    return move_dir(backup_dir, src_dir)
+  else:
+    return move_dir(src_dir, backup_dir)
+
 
 if os.geteuid() == 0:
   print 'ERROR: Please run as regular user'
   sys.exit(1)
 
-if os.path.exists(_CLIPS_DST):
-  if os.path.exists(_CLIPS_BAK):
-    print 'ERROR: Cannot overwrite:', _CLIPS_BAK
-    sys.exit(1)
-  os.rename(_CLIPS_DST, _CLIPS_BAK)
+if not backup('clips', False):
+  sys.exit(1)
+if not backup('playlists', False):
+  backup('clips', True)
+  sys.exit(1)
 
 shutil.rmtree('env')
 
-# subprocess.check_call(['pip', 'uninstall', 'pillow'])
-# subprocess.check_call(['pip', 'install', 'pillow'])
-
 subprocess.check_call(['make', 'develop'])
 
-if os.path.exists(_CLIPS_BAK):
-  os.rename(_CLIPS_BAK, _CLIPS_DST)
+backup('clips', True)
+backup('playlists', True)
 
 print '=== SUCCESS ==='
